Flexbox Update
Removed bootstrap, it was more complicated than necessary.
Added flexboxes to the navbar and home page. Look in navbar.css,
layout.css, home.css, and their respective php pages to see some
examples of flexboxes in use.
Added a flexbox to the forum and player bios pages. It's just a simple
div that wraps around the content and offsets it so it's not underneath
the navbar.
Added dark background.
Added a gradient to the navbar.