Ranking system and better MySQL queries
Changed the way databasetest.php queried the database so that it only
uses one query instead of two. It's harder to understand but it means
it's a lot more flexible.

Made an initial rankings page. It just ranks players by one of their
stats for now. Sai plans to add a column to each table that will store
their pre-calculated formula result so that they can be ranked. The
formula will have to be calculated every time a user enters data through
the GUI.