Projections
Refactored some of the graph code, it's still a little messy right now
(you have to manually specify the stats when creating the graph in JS).
You can now use a set number of points to project data (default is 3),
and projecting is as easy as calling a function.

Also converted the video for the home page.