Proper SQL Queries & Null Columns
In rankings.php I replaced the old queries with new ones that JOIN
several tables together. They are a little confusing but they allow for
null columns, so for example the hitting rankings now include all
hitters, regardless of whether or not they have any entries in the
player_strength table. I purposefully limited this query to only check
if there's a batspeed_id for a row, so that players who have entries in
player_strength but aren't hitters won't show up. In the fielding
rankings, I'm checking if any of the table ids aren't null, because I
want to include everybody who is either an outfielder, infielder,
catcher, or any combination of the three.

I also made the rank for hitting and speed a number that coaches will
understand: a player's place in regards to other players in the same
ranking. So that means if two players have the same score, they'll have
the same rank, just so it's fair.